requirements.txt
contourpy==1.3.1
cycler==0.12.1
digi-xbee==1.5.0
fonttools==4.55.3
invoke==2.2.0
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2024.2
watchdog==6.0.0
